+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.9.0] - 2023-12-04
 ### New
 - `User.referralLink` to retrieve the link Explorer which allow Explorer to get reward when friends join the Aira family (#98)
+- `PlatformClient.getPropertyValue` to retrieve properties from `UserPropertyName` (#98)
 
 ## [1.8.1] - 2023-11-06
 ### Changed